refactor(fieldmap): remove all data access to data fetch context
