+++ js/source/include/fieldmap/context.types.ts
@@ -1,5 +1,12 @@
-export interface FieldMapContextProps {
-	trialId: string;
+export interface FieldMapProps {
+    trialId: string;
+    trialStockType: string;
+    hasColAndRowNumbers: boolean;
+    hasSubplotEntries: boolean;
+    hasPlantEntries: boolean;
+    authToken?: string;
+}
 
+export interface FieldMapContextProps extends FieldMapProps {
 	children: React.ReactNode;
 }
