+++ js/source/include/fieldmap/contexts/LayoutConfigContext.tsx
@@ -5,27 +5,23 @@ export type PlotLayout = 'serpentine' | 'zigzag';
 export type ColorVar = 'parity' | 'germplasm' | 'block' | 'family_name' | 'cross_name';
 export type LabelVar = 'plot_number' | 'germplasm' | 'block' | 'family_name' | 'cross_name';
 
-interface LayoutConfigContextType {
+export interface LayoutConfigContextType {
     plotLayout: PlotLayout;
-    setPlotLayout: (val: PlotLayout) => void;
+    setPlotLayout: React.Dispatch<React.SetStateAction<PlotLayout>>;
     invertRows: boolean;
-    setInvertRows: (val: boolean) => void;
+    setInvertRows: React.Dispatch<React.SetStateAction<boolean>>;
     invertCols: boolean;
-    setInvertCols: (val: boolean) => void;
-    colorVar: ColorVar;
-    setColorVar: (val: ColorVar) => void;
-    labelVar: LabelVar;
-    setLabelVar: (val: LabelVar) => void;
-    labelSize: number;
-    setLabelSize: (val: number) => void;
+    setInvertCols: React.Dispatch<React.SetStateAction<boolean>>;
     topBorder: boolean;
-    setTopBorder: (val: boolean) => void;
+    setTopBorder: React.Dispatch<React.SetStateAction<boolean>>;
     bottomBorder: boolean;
-    setBottomBorder: (val: boolean) => void;
+    setBottomBorder: React.Dispatch<React.SetStateAction<boolean>>;
     leftBorder: boolean;
-    setLeftBorder: (val: boolean) => void;
+    setLeftBorder: React.Dispatch<React.SetStateAction<boolean>>;
     rightBorder: boolean;
-    setRightBorder: (val: boolean) => void;
+    setRightBorder: React.Dispatch<React.SetStateAction<boolean>>;
+    northArrowAngle: number;
+    setNorthArrowAngle: React.Dispatch<React.SetStateAction<number>>;
 }
 
 const LayoutConfigContext = createContext<LayoutConfigContextType | undefined>(undefined);
@@ -34,26 +30,22 @@ export const LayoutConfigProvider: React.FC<FieldMapContextProps> = ({ children 
     const [plotLayout, setPlotLayout] = useState<PlotLayout>('serpentine');
     const [invertRows, setInvertRows] = useState(false);
     const [invertCols, setInvertCols] = useState(false);
-    const [colorVar, setColorVar] = useState<ColorVar>('parity');
-    const [labelVar, setLabelVar] = useState<LabelVar>('plot_number');
-    const [labelSize, setLabelSize] = useState(10);
     const [topBorder, setTopBorder] = useState(false);
     const [bottomBorder, setBottomBorder] = useState(false);
     const [leftBorder, setLeftBorder] = useState(false);
     const [rightBorder, setRightBorder] = useState(false);
+    const [northArrowAngle, setNorthArrowAngle] = useState<number>(0);
 
     return (
         <LayoutConfigContext.Provider value={{
             plotLayout, setPlotLayout,
             invertRows, setInvertRows,
             invertCols, setInvertCols,
-            colorVar, setColorVar,
-            labelVar, setLabelVar,
-            labelSize, setLabelSize,
             topBorder, setTopBorder,
             bottomBorder, setBottomBorder,
             leftBorder, setLeftBorder,
-            rightBorder, setRightBorder
+            rightBorder, setRightBorder,
+            northArrowAngle, setNorthArrowAngle
         }}>
             {children}
         </LayoutConfigContext.Provider>
