+++ js/source/include/fieldmap/modals/DimensionsModal.tsx
@@ -1,36 +1,38 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { AccessionAutocomplete } from '../components/AccessionAutocomplete';
+import { useModals } from '../contexts/ModalsContext';
+import { useDataFetch } from '../contexts/DataFetchContext';
 
 interface DimensionsModalProps {
-    show: boolean;
-    onClose: () => void;
-    dimRowsInput: string;
-    setDimRowsInput: (v: string) => void;
-    dimColsInput: string;
-    setDimColsInput: (v: string) => void;
-    fillerAccessionInput: string;
-    setFillerAccessionInput: (v: string) => void;
-    onApply: () => void;
 }
 
-export const DimensionsModal: React.FC<DimensionsModalProps> = ({
-    show,
-    onClose,
-    dimRowsInput,
-    setDimRowsInput,
-    dimColsInput,
-    setDimColsInput,
-    fillerAccessionInput,
-    setFillerAccessionInput,
-    onApply
-}) => {
+export const DimensionsModal: React.FC<DimensionsModalProps> = ({}) => {
+    const {
+        showDimDialog: show,
+        setShowDimDialog: setShow,
+    } = useModals();
+
+    const [dimRowsInput, setDimRowsInput] = useState('');
+    const [dimColsInput, setDimColsInput] = useState('');
+    const [fillerAccessionInput, setFillerAccessionInput] = useState('');
+
+    const {
+        applyDimensions
+    } = useDataFetch();
+
     if (!show) return null;
+
+    const handleApplyDimensions = async () => {
+        await applyDimensions(dimRowsInput, dimColsInput, fillerAccessionInput);
+        setShow(false);
+    };
+
     return (
         <div className="modal show tw:block tw:bg-black/50">
             <div className="modal-dialog">
                 <div className="modal-content">
                     <div className="modal-header">
-                        <button type="button" className="close" onClick={onClose}>&times;</button>
+                        <button type="button" className="close" onClick={() => setShow(false)}>&times;</button>
                         <h4 className="modal-title">Change Layout Dimensions</h4>
                     </div>
                     <div className="modal-body">
@@ -48,8 +50,8 @@ export const DimensionsModal: React.FC<DimensionsModalProps> = ({
                         </div>
                     </div>
                     <div className="modal-footer">
-                        <button className="btn btn-default" onClick={onClose}>Cancel</button>
-                        <button className="btn btn-primary" onClick={onApply}>Apply</button>
+                        <button className="btn btn-default" onClick={() => setShow(false)}>Cancel</button>
+                        <button className="btn btn-primary" onClick={handleApplyDimensions}>Apply</button>
                     </div>
                 </div>
             </div>
