+++ js/source/include/fieldmap/modals/DownloadCSVModal.tsx
@@ -1,37 +1,46 @@
-import React from 'react';
-
-export interface CSVDownloadOpts {
-    accession: boolean;
-    obsUnit: boolean;
-    seedlot: boolean;
-    plotId: boolean;
-    plotNum: boolean;
-    familyName: boolean;
-    crossName: boolean;
-}
+import React, { useState } from 'react';
+import { useModals } from '../contexts/ModalsContext';
+import { CsvDownloadOpts } from '../model.types';
+import { downloadLayoutCSV } from '../utils/downloadCsv';
+import { useLayoutConfig } from '../contexts/LayoutConfigContext';
+import { usePlotGrid } from '../contexts/PlotGridContext';
 
 interface DownloadCSVModalProps {
-    show: boolean;
-    onClose: () => void;
-    csvDownloadOpts: CSVDownloadOpts;
-    setCsvDownloadOpts: React.Dispatch<React.SetStateAction<CSVDownloadOpts>>;
-    onDownload: () => void;
+    trialId: string;
 }
 
-export const DownloadCSVModal: React.FC<DownloadCSVModalProps> = ({
-    show,
-    onClose,
-    csvDownloadOpts,
-    setCsvDownloadOpts,
-    onDownload
-}) => {
+export const DownloadCSVModal: React.FC<DownloadCSVModalProps> = ({ trialId }) => {
+    const {
+        showDownloadCSVModal: show,
+        setShowDownloadCSVModal: setShow
+    } = useModals();
+
     if (!show) return null;
+
+    const { plotList, bounds } = usePlotGrid();
+    const { invertCols, invertRows } = useLayoutConfig();
+
+    const [csvDownloadOpts, setCsvDownloadOpts] = useState<CsvDownloadOpts>({
+        accession: true,
+        obsUnit: false,
+        seedlot: false,
+        plotId: false,
+        plotNum: false,
+        familyName: false,
+        crossName: false,
+    });
+
+    const handleDownload = () => {
+        downloadLayoutCSV(trialId, bounds, plotList, invertCols, invertRows, csvDownloadOpts);
+        setShow(false);
+    };
+
     return (
         <div className="modal show tw:block tw:bg-black/50">
             <div className="modal-dialog">
                 <div className="modal-content">
                     <div className="modal-header">
-                        <button type="button" className="close" onClick={onClose}>&times;</button>
+                        <button type="button" className="close" onClick={() => setShow(false)}>&times;</button>
                         <h4 className="modal-title">Download Spatial Layout Customizer</h4>
                     </div>
                     <div className="modal-body">
@@ -58,8 +67,8 @@ export const DownloadCSVModal: React.FC<DownloadCSVModalProps> = ({
                         </div>
                     </div>
                     <div className="modal-footer">
-                        <button className="btn btn-default" onClick={onClose}>Close</button>
-                        <button className="btn btn-primary" onClick={onDownload}>Download CSV</button>
+                        <button className="btn btn-default" onClick={() => setShow(false)}>Close</button>
+                        <button className="btn btn-primary" onClick={handleDownload}>Download CSV</button>
                     </div>
                 </div>
             </div>
