+++ js/source/include/fieldmap/modals/SuppressPhenotypeModal.tsx
@@ -1,27 +1,46 @@
-import React from 'react';
+import React, { useMemo } from 'react';
+import { useModals } from '../contexts/ModalsContext';
+import { useView } from '../contexts/ViewContext';
+import { useHeatmap } from '../contexts/HeatmapContext';
+import { useDataFetch } from '../contexts/DataFetchContext';
 
 interface SuppressPhenotypeModalProps {
-    show: boolean;
-    onClose: () => void;
-    plotName: string;
-    phenotypeValue: number | string | undefined;
-    onSuppress: () => void;
 }
 
-export const SuppressPhenotypeModal: React.FC<SuppressPhenotypeModalProps> = ({
-    show,
-    onClose,
-    plotName,
-    phenotypeValue,
-    onSuppress
-}) => {
+export const SuppressPhenotypeModal: React.FC<SuppressPhenotypeModalProps> = ({ }) => {
+    const {
+        showSuppressModal: show,
+        setShowSuppressModal: setShow
+    } = useModals();
+
     if (!show) return null;
+
+    const {
+        selectedPlot,
+    } = useView();
+
+    const {
+        heatmapData
+    } = useHeatmap();
+
+    const {
+        handleSuppressPhenotype
+    } = useDataFetch();
+
+    const plotName = useMemo(() => {
+        return selectedPlot?.observationUnitName || '';
+    }, [selectedPlot]);
+
+    const phenotypeValue = useMemo(() => {
+        return selectedPlot ? heatmapData[selectedPlot.observationUnitDbId || '']?.val : undefined;
+    }, [selectedPlot, heatmapData]);
+
     return (
         <div className="modal show tw:block tw:bg-black/50">
             <div className="modal-dialog">
                 <div className="modal-content">
                     <div className="modal-header">
-                        <button type="button" className="close" onClick={onClose}>&times;</button>
+                        <button type="button" className="close" onClick={() => setShow(false)}>&times;</button>
                         <h4 className="modal-title">Suppress Plot Phenotype Measurement</h4>
                     </div>
                     <div className="modal-body">
@@ -30,8 +49,8 @@ export const SuppressPhenotypeModal: React.FC<SuppressPhenotypeModalProps> = ({
                         <div><strong>Phenotype Value:</strong> {phenotypeValue}</div>
                     </div>
                     <div className="modal-footer">
-                        <button className="btn btn-default" onClick={onClose}>Close</button>
-                        <button className="btn btn-danger" onClick={onSuppress}>Suppress Phenotype</button>
+                        <button className="btn btn-default" onClick={() => setShow(false)}>Close</button>
+                        <button className="btn btn-danger" onClick={handleSuppressPhenotype}>Suppress Phenotype</button>
                     </div>
                 </div>
             </div>
